Fix tab drag interactive UI test failures
When there is a grab state due to window drag, the completion events may
not come through the wayland input event handler. So listen for window
drag signals explicitly to notify completion of motion and release
events in that state. This required adding two additional signals,
"update-move-done" and "update-resize-done" to MetaWindowDrag class.

Also, handle the case where a move relative to a window maybe requested
immediately after window creation, resulting in unexpected positioning
of the window. Add a check to wait for all window effects to complete
before initiating the request.

Change-Id: I41910ff64f02552828c7e1f9ff87becf428a1226